+++ objc2-foundation/src/process_info.rs
@@ -1,3 +1,6 @@
+use core::fmt;
+use core::panic::{RefUnwindSafe, UnwindSafe};
+
 use objc2::msg_send_id;
 use objc2::rc::{Id, Shared};
 
@@ -7,25 +10,51 @@ extern_class! {
     /// A collection of information about the current process.
     ///
     /// See [Apple's documentation](https://developer.apple.com/documentation/foundation/nsprocessinfo?language=objc).
-    #[derive(Debug, PartialEq, Eq, Hash)]
+    #[derive(PartialEq, Eq, Hash)]
     unsafe pub struct NSProcessInfo: NSObject;
-
-    // TODO: This contains a lot more important functionality!
 }
 
-// The documentation explicitly states:
+// SAFETY: The documentation explicitly states:
 // > NSProcessInfo is thread-safe in macOS 10.7 and later.
 unsafe impl Send for NSProcessInfo {}
 unsafe impl Sync for NSProcessInfo {}
 
+impl UnwindSafe for NSProcessInfo {}
+impl RefUnwindSafe for NSProcessInfo {}
+
 impl NSProcessInfo {
     pub fn process_info() -> Id<NSProcessInfo, Shared> {
-        // currentThread is @property(strong), what does that mean?
-        // TODO: Always available?
         unsafe { msg_send_id![Self::class(), processInfo].unwrap() }
     }
 
     pub fn process_name(&self) -> Id<NSString, Shared> {
         unsafe { msg_send_id![self, processName].unwrap() }
+    }
+
+    // TODO: This contains a lot more important functionality!
+}
+
+impl fmt::Debug for NSProcessInfo {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("NSProcessInfo")
+            .field("process_name", &self.process_name())
+            .finish_non_exhaustive()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use alloc::format;
+
+    use super::*;
+
+    #[test]
+    fn test_debug() {
+        let info = NSProcessInfo::process_info();
+        let expected = format!(
+            "NSProcessInfo {{ process_name: {:?}, .. }}",
+            info.process_name()
+        );
+        assert_eq!(format!("{:?}", info), expected);
     }
 }