+++ objc2-foundation/CHANGELOG.md
@@ -6,6 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased - YYYY-MM-DD
 
+### Added
+* Implement `UnwindSafe` and `RefUnwindSafe` for all objects.
+* Implemented `IntoIterator` for references to `NSArray`, `NSMutableArray`,
+  `NSData` and `NSMutableData`.
+* Implemented `Extend` for `NSMutableArray`.
+* Add extra `Extend<&u8>` impl for `NSMutableData`.
+
+### Fixed
+* Made `Debug` impls for all objects print something useful.
+
+### Removed
+* `NSObject::hash_code`, `NSObject::is_equal` and `NSObject::description` in
+  favour of just having the trait implementations `Hash`, `PartiqalEq` and
+  `Debug`.
+
 
 ## 0.2.0-alpha.6 - 2022-07-19